After investigations, finalise security cfg for Oauth2-api
If you are not logged in (or have a bearer token) in /bearer-api,
you will need to login.  But, if you have a valid token you can use it.

The order of fiewwalls and access_control's are important. More specific first!
The last access-control that matches the path will be used.