Don't be so keen to redirect back - we need to control when this happens
We are storing it elsewhere, just below in starlingOAuthSession(),
and we keep that for the post-purchase welcome screen to link-back to
Starling. That gives them API access, via the oauth/v2/auth to send
the state & a code to upgrade to an access-code token.

We may not need all the allowed redirect URLs, but we will need the
sandbox and then prod/live oauth endpoints for Starling for the final
redirect back.