Translating the sitemap to a little more human language
The models etc, have an added 'description' in the sitemap 'Entry' object.
Other URLs are translated in a twig extension. If any are not matched,
the URL falls through as a no-op and is displayed as before.

The model description has seperate link-descriptions for the plain make/model,
and any with a memory-size that are also listed - they go to different URLs.